requirements.txt
autopep8==1.3.5
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
contextlib2==0.5.5
diff-match-patch==20121119
Django==2.1
django-import-export==1.0.1
djangorestframework==3.8.2
et-xmlfile==1.0.1
fake-useragent==0.1.10
funcsigs==1.0.2
future==0.16.0
Google-Search-API==1.1.14
idna==2.6
jdcal==1.4
mock==2.0.0
multidict==4.3.1
odfpy==1.3.6
openpyxl==2.5.5
pbr==4.0.2
psycopg2==2.7.3.2
pycodestyle==2.4.0
pytz==2018.5
PyYAML==3.12
requests==2.19.1
selenium==2.53.6
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
vcrpy==1.11.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
yarl==1.2.6
